Updated authentication and registration. Now using crypto Spring Security. Having trouble adding new users to the database. I think it's an issue with the DataSeeder.

+++ PlatePlanner/src/main/java/org/launchcode/PlatePlanner/service/MealPlanService.java
@@ -19,7 +19,7 @@ public class MealPlanService {
 
     //Users can create a meal plan so long as the name is unique
 
-    public MealPlan createMealPlan(Long userId, String mealPlanName) {
+    public MealPlan createMealPlan(Integer userId, String mealPlanName) {
         Optional<User> optionalUser = userRepository.findById(userId);
 
         if (optionalUser.isEmpty()) {
